processed the review code
