+++ src/styles.css
@@ -48,7 +48,6 @@ body {
 }
 
 /*background_module*/
-
 .background-module {
     width: 180px;
     padding: 20px;
@@ -65,11 +64,8 @@ body {
     display: flex;
 }
 
-.background-module__color span {
-    opacity: 0;
-}
-
 .background-module__btn {
+    width: 180px;
     margin-bottom: 20px;
     padding: 10px;
     border: 1px solid transparent;
@@ -77,7 +73,7 @@ body {
     outline: transparent;
     background-color: rgb(229, 228, 226);
     cursor: pointer;
-    transition: all 0.2s ease-in;
+    transition: all .2s ease-in;
 }
 
 .background-module__btn:hover {
@@ -85,10 +81,10 @@ body {
 }
 
 .background-module__btn:focus {
-    box-shadow: 0px 0px 0px 3px rgba(0, 0, 0, 0.25);
+    box-shadow: 0 0 0 3px rgba(0, 0, 0, .25);
 }
 
 .background-module__btn:active {
-    color: rgba(255, 255, 255, 0.5);
-    box-shadow: inset 0px 4px 4px rgba(0, 0, 0, 0.15);
+    color: rgba(255, 255, 255, .5);
+    box-shadow: inset 0 4px 4px rgba(0, 0, 0, .15);
 }
